requirements.txt
opentelemetry-distro[otlp]==0.58b0
opentelemetry-instrumentation==0.58b0
# Removendo versão específica - deixar o distro resolver
# opentelemetry-exporter-otlp será instalado pelo distro[otlp]
opentelemetry-instrumentation-django==0.58b0
opentelemetry-instrumentation-psycopg2==0.58b0
opentelemetry-instrumentation-requests==0.58b0
opentelemetry-instrumentation-logging==0.58b0
# Para instrumentação Kafka (se usando kafka-python)
opentelemetry-instrumentation-kafka-python==0.58b0
# Para métricas HTTP automaticas
opentelemetry-instrumentation-wsgi==0.58b0
opentelemetry-instrumentation-asgi==0.58b0